update return values of some methods in method helper interface to accomodate the database returning longs instead of ints

+++ project_phase3/features/src_app/app/src/main/java/com/b07/database/helper/DatabasePlatformHelper.java
@@ -44,9 +44,9 @@ public interface DatabasePlatformHelper {
   public int insertAccountLine(int accountId, int itemId, int quantity)
           throws DatabaseInsertException, SQLException;
 
-  public int insertDiscountType(String name) throws DatabaseInsertException, SQLException;
+  public long insertDiscountType(String name) throws DatabaseInsertException, SQLException;
 
-  public int insertCoupon(int itemId, int uses, String type, BigDecimal discount, String code)
+  public long insertCoupon(int itemId, int uses, String type, BigDecimal discount, String code)
           throws DatabaseInsertException, SQLException;
 
   //UpdateHelper
